+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityApplicationModel.java
@@ -26,7 +26,7 @@ import is.codion.framework.model.DefaultEntityApplicationModel;
  * A Swing implementation of {@link is.codion.framework.model.EntityApplicationModel}
  */
 public class SwingEntityApplicationModel
-				extends DefaultEntityApplicationModel<SwingEntityModel, SwingEntityEditModel, SwingEntityTableModel> {
+				extends DefaultEntityApplicationModel<SwingEntityEditModel, SwingEntityTableModel> {
 
 	/**
 	 * Instantiates a new {@link SwingEntityApplicationModel}