+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityModel.java
@@ -30,7 +30,7 @@ import static java.util.Objects.requireNonNull;
 /**
  * A Swing implementation of {@link is.codion.framework.model.EntityModel}
  */
-public class SwingEntityModel extends AbstractEntityModel<SwingEntityModel, SwingEntityEditModel, SwingEntityTableModel> {
+public class SwingEntityModel extends AbstractEntityModel<SwingEntityEditModel, SwingEntityTableModel> {
 
 	/**
 	 * Instantiates a new SwingEntityModel with default SwingEntityEditModel and SwingEntityTableModel implementations.